+++ src/comparatrain/Evaluable.java
@@ -5,8 +5,12 @@ package comparatrain;
 
 /**
  * @author Vivien Galuchot - Vincent Hernandez
- *
+ * Evaluable, objet contenant un ensemble de criteres élémentaires
  */
 public interface Evaluable {
-	int eval(Preference pref);
+	/**
+	 * @param pref preferences permetant d'effectuer l'évaluation
+	 * @return un réel entre 0 et 1, 1 etant le plus proche de preference
+	 */
+	double eval(Preference pref);
 }
